Double bookings - Race condition
There's a racing condition in which if two or more users reserve the same resource at the exact same time then only the last one gets displays even though both suceeded in reserving the item, check here for more details:
https://forums.bookedscheduler.com/viewtopic.php?f=3&t=423

This is not the ideal solution and I hope it will only be temporary, but I rather have this than having no fix at all. It's still possible to have a race condition but it should be more improbable to happen.