BUGFIX - API: reflect qry param in GetAvailability
With GetAvailability(resourceId = null) the GET query parameter is not reflected in the response. Maybe this was intermittently set since there is an alternative route for GetAvailability that uses a static URI instead of a query parameter for resourceId. See /Web/Services/index.php . This fix prefers the route with explicit query params for resourceId and dateTime over the static route. Hence, the alternative route should be disabled with a subsequent change.

Please feel free to correct me if I misunderstood the function signature or if there is a way to enable both routes. I have tried but the Slim framework seems not to allow alternative signatures on the same function.