feat: new algorithm for fast loading reservations
Add a new algorithm for fast loading reservations in the schedule.

Marking it as "experimental" and users will have to explicitly enable
it.

Limitations:
  * Only supports the ScheduleStandard style. Otherwise it will use
    the default legacy algorithm.
  * Has not been setup for concurrent reservations. So behavior is
    undefined for concurrent reservation.

The performance gain is very significant when loading larger
schedules.  On a one day schedule with 332 resources. The legacy
algorithm took 23 seconds to "Load Reservations". While the new
algorithm took 0.6 seconds. "Load Reservations" is the time for the
renderEvents() method to run. It does not count the time for the web
page to be fetched from the server.

The more resources / days in the schedule the larger the difference in
timing becomes.

Closes: #370